requirements.txt
aiodns==3.2.0
aiohappyeyeballs==2.4.3
aiohttp==3.10.10
aiohttp-retry==2.8.3
aiosignal==1.3.1
annotated-types==0.7.0
anyio==4.6.2.post1
appnope==0.1.4
asttokens==2.4.1
attrs==24.2.0
backoff==2.2.1
bcrypt==4.2.0
beautifulsoup4==4.12.3
boto3==1.35.43
botocore==1.35.43
Brotli==1.1.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
comm==0.2.2
contourpy==1.3.0
crackpy @ git+https://github.com/Rievil/CrackPy@e64ce0b5914852c6878c799b938db2224a46e912
cryptography==43.0.1
cycler==0.12.1
debugpy==1.8.6
decorator==5.1.1
dnspython==2.7.0
efficientnet_pytorch==0.7.1
email_validator==2.2.0
executing==2.1.0
fastapi==0.115.2
fastapi-cli==0.0.5
filelock==3.16.1
fonttools==4.54.1
frozenlist==1.4.1
fsspec==2024.9.0
gdown==5.2.0
h11==0.14.0
httpcore==1.0.6
httptools==0.6.4
httpx==0.27.2
huggingface-hub==0.25.1
idna==3.10
imageio==2.35.1
inquirerpy==0.3.4
IProgress==0.4
ipykernel==6.29.5
ipython==8.27.0
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
jmespath==1.0.1
joblib==1.4.2
jupyter_client==8.6.3
jupyter_core==5.7.2
kiwisolver==1.4.7
lazy_loader==0.4
llvmlite==0.43.0
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.9.2
matplotlib-inline==0.1.7
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
munch==4.0.0
nest-asyncio==1.6.0
networkx==3.3
numba==0.60.0
numpy==2.0.2
opencv-python==4.10.0.84
orjson==3.10.7
packaging==24.1
pandas==2.2.3
paramiko==3.5.0
parso==0.8.4
pexpect==4.9.0
pfzy==0.3.4
pillow==10.4.0
platformdirs==4.3.6
pretrainedmodels==0.7.4
prettytable==3.11.0
prompt_toolkit==3.0.48
propcache==0.2.0
psutil==6.0.0
ptyprocess==0.7.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pycares==4.4.0
pycparser==2.22
pydantic==2.9.2
pydantic-extra-types==2.9.0
pydantic-settings==2.6.0
pydantic_core==2.23.4
Pygments==2.18.0
PyNaCl==1.5.0
pyparsing==3.1.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-multipart==0.0.12
pytz==2024.2
PyYAML==6.0.2
pyzmq==26.2.0
requests==2.32.3
rich==13.9.2
runpod==1.7.3
s3transfer==0.10.3
safetensors==0.4.5
scikit-image==0.24.0
scikit-learn==1.5.2
scipy==1.14.1
segmentation-models-pytorch==0.3.4
setuptools==75.1.0
shellingham==1.5.4
six==1.16.0
sknw==0.15
sniffio==1.3.1
soupsieve==2.6
stack-data==0.6.3
starlette==0.40.0
sympy==1.13.3
threadpoolctl==3.5.0
tifffile==2024.9.20
timm==0.9.7
tomli==2.0.2
tomlkit==0.13.2
torch==2.4.1
torchvision==0.19.1
tornado>=6.4.2
tqdm==4.66.5
tqdm-loggable==0.2
traitlets==5.14.3
typer==0.12.5
typing_extensions==4.12.2
tzdata==2024.2
ujson==5.10.0
urllib3==2.2.3
uvicorn==0.32.0
uvloop==0.21.0
Wand==0.6.13
watchdog==5.0.3
watchfiles==0.24.0
wcwidth==0.2.13
websockets==13.1
yarl==1.15.4
